project/requirements.txt
# flask
# gunicorn
Flask[async]==2.0.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
gunicorn==20.1.0
click==8.0.3
itsdangerous==2.0.1
Jinja2==3.0.3
Werkzeug==2.0.2
jmespath==0.10.0
s3transfer==0.5.0
boto3==1.20.27
botocore==1.23.27
redis==4.1.2
Deprecated==1.2.13
packaging==21.3
rq==1.10.1
# rq-dashboard==0.6.1
rq-scheduler==0.11.0
stripe==2.64.0
requests==2.27.0
python-dateutil==2.8.2
python-dotenv==0.19.2
croniter==1.2.0
bcrypt
